% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_drawio.R
\name{generate_drawio}
\alias{generate_drawio}
\title{Generate a drawio document based on a template and metadata}
\usage{
generate_drawio(template, meta, output = NULL,
  drawio_path = "C:/Program Files/draw.io/draw.io.exe")
}
\arguments{
\item{template}{The path to the drawio template}

\item{meta}{A liste containing metadata}

\item{output}{A path for the export file (allowed formats : png, jpg, svg, pdf)}

\item{drawio_path}{The path to draw.io executable}
}
\description{
Your template must contains placeholders like %placeholder%
(see: https://desk.draw.io/support/solutions/articles/16000051979-how-to-work-with-placeholders-).
Placeholders must be activated for your objects (CTRL+M).
This function will populate your placeholders based on the meta argument and the template file,
and will export the resulting chart in a desired format.
}
\details{
Draw.io desktop version is required to generate exports.
It is assumed to be 'C:/Program Files/draw.io/draw.io.exe' but it can be changed with drawio_path argument
}
\examples{
\dontrun{
meta <- list('color'='red', 'age'=75)
generate_drawio('path/to/template.drawio', meta, 'path/to/export.png')
}

}
\author{
Jeremy Pasco
}
